requirements.txt
flask
dash
dash_table
dash-bootstrap-components
plotly
pandas
python-dotenv
langchain
langchain_community
ollama
pyjwt
python-jose[cryptography]
fastapi